Implemented RocketListViewModel for RocketListFragment.kt and structured project architecture due to mvvm rules
