+++ app/src/main/AndroidManifest.xml
@@ -14,14 +14,14 @@
         android:theme="@style/Theme.DragonX">
         <activity android:name=".RocketDetails"
             android:label="@string/title_activity_rocket_details"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName=".view.activities.MainActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity"/>
+                android:value=".view.activities.MainActivity"/>
         </activity>
 
         <activity
-            android:name=".SplashActivity"
+            android:name=".view.activities.SplashActivity"
             android:theme="@style/Theme.DragonX.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -29,7 +29,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivity" />
+        <activity android:name=".view.activities.MainActivity" />
 
         <meta-data
             android:name="preloaded_fonts"
