+++ app/src/main/java/com/example/dragonx/view/activities/MainActivity.kt
@@ -1,0 +1,17 @@
+
+package com.example.dragonx.view.activities
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.dragonx.R
+
+@Suppress("DEPRECATION", "BlockingMethodInNonBlockingContext")
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+
+}
+
