+++ app/src/main/java/com/example/dragonx/view/fragments/RocketDetailsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.dragonx
+package com.example.dragonx.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,6 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import com.denzcoskun.imageslider.constants.ScaleTypes
 import com.denzcoskun.imageslider.models.SlideModel
+import com.example.dragonx.JsonDetailsParser
+import com.example.dragonx.R
 import com.example.dragonx.models.Rocket
 import kotlinx.android.synthetic.main.fragment_rocket_details.view.*
 import kotlinx.coroutines.CoroutineScope
