+++ app/src/main/res/layout/fragment_rocket_details.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".RocketDetailsFragment">
+    tools:context=".view.fragments.RocketDetailsFragment">
 
     <ScrollView
         android:layout_width="match_parent"
